Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo>=4.9,<4.10
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-multipart==0.0.9